requirements.txt
apturl==0.5.2
asgiref==3.8.1
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2024.8.30
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
diff-match-patch==20241021
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.1
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-extensions==3.2.3
django-import-export==4.2.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.1.0
duplicity==0.8.21
et_xmlfile==2.0.0
fasteners==0.14.1
filelock==3.16.1
future==0.18.2
git-filter-repo==2.45.0
gpg==1.16.0
gunicorn==23.0.0
gyp==0.1
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.3.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
mysql-connector-python==9.0.0
netifaces==0.11.0
numpy==2.1.2
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
paramiko==2.9.3
pexpect==4.8.0
pillow==10.4.0
pipenv==2024.3.0
platformdirs==4.3.6
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.9.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.32.3
requests-toolbelt==0.9.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.1
systemd-python==234
tablib==3.7.0
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
usb-creator==0.3.7
virtualenv==20.27.1
virtualenv-clone==0.3.0
wadllib==1.3.6
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zipp==1.0.0
